jwt token created and authentication

+++ src/routes/memories.ts
@@ -3,8 +3,19 @@ import { z } from 'zod' // for validation
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify() //verify the jwt token before all these routes
+  })
+
+  app.get('/memories', async (request) => {
+    //await request.jwtVerify() //verify the jwt token only here
+
+    //typescrit dont recognize request.user.variables automatically, i have to create a auth.d.ts file (it's present in https://github.com/fastify/fastify-jwt)
+
     const memories = await prisma.memory.findMany({
+      where:{
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -17,13 +28,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
       }
     })
   })
-  /*    const users = await prisma.user.findMany({
-      select: { pick just the name
-        name: true,
-      },
-    }) */
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     // const { id } = request.params
 
     const paramsSchema = z.object({
@@ -37,6 +43,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -54,14 +64,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '391342b5-040a-445d-b43b-3964d1e0588b', // remember to get the user id
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -75,7 +85,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if(memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -88,12 +108,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where:{
+        id,
+      }
+    })
+
+    if(memory.id !== request.user.sub){
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
